ACI/requirements.txt
fastapi
uvicorn
tensorflow
pillow
opencv-python-headless
python-dotenv
matplotlib
networkx
pandas
scikit-image
scikit-learn
patchify
scipy
skan
jinja2
apscheduler
sqlalchemy
sqlalchemy_utils
databases
celery[redis]
redis
aiofiles
numpy
fastapi-mail